Replaced 3rd party jQuery JSONRPC plug in with my own.
Had some problems with the original plug in. Looked inside and could not figure out what is what. On top of the fact that it was written for JSONPRC v.2.0, and our server RPC does v.1.0, you had convoluted object instantiation all over the place. I am guessing the author is either a Ruby or Java enthusiast. Instead of spending time to figure out how to fix it, just looked at the spec and wrote my own. 40 lines of code. 50 lines of comments.