Adding forced text wrap for long path labels.
Path Crumbs is a long string, which has no interupted space in it, unless some file name has a space in it. This means that it does not wrap by default.

Found a relatively reliable way to make the long strings wrap here: http://bytes.com/topic/html-css/answers/627827-how-wrap-text-css