Fixed -Invalid directory listing.- error when viewing fresh, empty repos.
Rather interesting... Because i use auto-create on push, I never had a completely empty, new repo anywhere on the paths. It's always, a repo with stuff in it or nothing. A user reported a -"Invalid directory listing."- problem during viewing of new repos that he likely created by hand. Now that use-case is accounted for in the UI and tests.